Major bug, in which std. dev. is incorrectly 0, fixed
This bug (now fixed) causes the computed std. dev. to become 0 and the
t-test to produce the wrong conclusion whenever the std. dev. is
calculated from a list of values.

+++ src/pfnguyen/statlogic/ttest/OneSampleTTest.java
@@ -185,7 +185,7 @@ public class OneSampleTTest {
      * @return    the sample variance
      */
     private double calcSampleVariance(ArrayList<Double> x) {
-        double oneOverNAdjusted = 1 / (x.size() - 1);
+        double oneOverNAdjusted = (double) 1 / (x.size() - 1);
         double sumOfDifference;
         double squaredDifference;
         double sumOfSquaredDifference = 1; // This is bad
